requirements.txt
# UI & Web Framework
streamlit

# Deep Learning & Computer Vision
torch
torchvision
facenet-pytorch
timm
Pillow
opencv-python-headless

# Data Analysis & Visualization
numpy
pandas
plotly

# Utilities
reportlab
gdown